Refactor #30 - 폴더 리팩토링
