+++ src/main/java/com/maruhxn/boardserver/auth/common/AjaxAuthenticationToken.java
@@ -1,4 +1,4 @@
-package com.maruhxn.boardserver.auth;
+package com.maruhxn.boardserver.auth.common;
 
 import org.springframework.security.authentication.AbstractAuthenticationToken;
 import org.springframework.security.core.GrantedAuthority;
