+++ src/main/java/com/maruhxn/boardserver/auth/provider/AjaxAuthenticationProvider.java
@@ -1,5 +1,7 @@
-package com.maruhxn.boardserver.auth;
+package com.maruhxn.boardserver.auth.provider;
 
+import com.maruhxn.boardserver.auth.common.AccountContext;
+import com.maruhxn.boardserver.auth.common.AjaxAuthenticationToken;
 import com.maruhxn.boardserver.common.exception.ErrorCode;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationProvider;
